+++ src/components/chat/qna/chatuser/UserChatRoom.tsx
@@ -1,40 +1,13 @@
-import { useEffect, useRef, useState } from 'react';
-import { supabase } from '../../../../api/supabase.api';
-import { useAuth } from '../../../../contexts/auth.context';
+import { useEffect, useRef } from 'react';
+import { MessageProps } from '../../../../types/types';
 import * as St from './UserChatRoom.styled';
 
-interface Message {
-  created_at: string;
-  content: string;
-  sender_id: string;
-  message_type: string;
-  id: number;
-}
-
-const ChatLog = () => {
-  const auth = useAuth();
-  const [messages, setMessages] = useState<Message[]>([]);
+const ChatLog = ({ messages }: MessageProps) => {
   const messageEndRef = useRef<HTMLDivElement | null>(null);
 
   useEffect(() => {
-    if (!auth.session) return;
-
-    getQnaLog(auth.session.user.id);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     messageEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-  }, [auth.session, messages]);
-
-  //qna table 가져오는 함수
-  const getQnaLog = async (roomId: string) => {
-    if (!auth.session) return;
-    const response = await supabase.from('qna').select('*').eq('room_id', roomId);
-    const result = response.data;
-    if (result) {
-      setMessages(result);
-    } else {
-      setMessages([]);
-    }
-  };
+  }, [messages]);
 
   return (
     <St.Container>
